requirements.txt
pandas
numpy
requests
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
torch
streamlit
joblib
tensorflow 
matplotlib 
plotly
fastapi 
uvicorn 
jinja2 


